[ ] 编辑角色的时候，选择上级角色会出现错误 [ ] 删除管理员限制只能删除下级管理员

+++ app/Application/Admin/Controller/UserController.php
@@ -38,6 +38,12 @@ class UserController extends AbstractController
             return $this->returnErrorJson('找不到该记录');
         }
 
+        $role_ids = AdminRole::getChildRoleIds(AdminUserService::getInstance()
+            ->getAdminUserRoleId());
+        if (!in_array($admin_user->role_id, $role_ids)) {
+            return $this->returnErrorJson("没有权限操作");
+        }
+
         return $admin_user->delete() ? $this->returnSuccessJson() : $this->returnErrorJson();
     }
